refactor MailboxLock into LocalMailboxLock and LocalMailboxLockFactory
