+++ store/src/java/com/zimbra/cs/mailbox/LocalMailboxLockFactory.java
@@ -1,7 +1,11 @@
 package com.zimbra.cs.mailbox;
 
+import java.util.EmptyStackException;
 import java.util.Stack;
+import java.util.concurrent.TimeUnit;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.zimbra.common.localconfig.LC;
 import org.apache.curator.framework.recipes.locks.InterProcessSemaphoreMutex;
 
 import com.zimbra.common.localconfig.DebugConfig;
@@ -16,35 +20,40 @@ import com.zimbra.cs.zookeeper.CuratorManager;
 
 
 public class LocalMailboxLockFactory implements MailboxLockFactory {
-	private final ZLock zLock = DebugConfig.debugMailboxLock ? new DebugZLock() : new ZLock();
+    private final ZLock zLock = DebugConfig.debugMailboxLock ? new DebugZLock() : new ZLock();
     private InterProcessSemaphoreMutex dLock = null;
-    private Mailbox mbox;
+    private final Stack<Boolean> lockStack = new Stack<>();
+    private final Mailbox mailbox;
+    //for sanity checking, we keep list of read locks. the first time caller obtains write lock they must not already own read lock
+    //states - no lock, read lock only, write lock only
+    private final ThreadLocal<Boolean> assertReadLocks = new ThreadLocal<>();
 
-    public LocalMailboxLockFactory(final String id, final Mailbox mbox) {
-    	if (Zimbra.isAlwaysOn()) {
+    public LocalMailboxLockFactory(final Mailbox mailbox) {
+        this.mailbox = mailbox;
+
+        if (Zimbra.isAlwaysOn()) {
             try {
-                dLock = CuratorManager.getInstance().createLock(id);
+                dLock = CuratorManager.getInstance().createLock(this.mailbox.getAccountId());
             } catch (ServiceException se) {
                 ZimbraLog.mailbox.error("could not initialize distributed lock", se);
             }
         }
-        this.mbox = mbox;
     }
 
     @Override
     public MailboxLock readLock() {
-        return new LocalMailboxLock (zLock.readLock(), false,zLock,dLock,mbox);
+        return new LocalMailboxLock(false);
     }
 
     @Override
     public MailboxLock writeLock() {
-    	 return new LocalMailboxLock (zLock.writeLock(), true,zLock,dLock,mbox);
+        return new LocalMailboxLock(true);
     }
 
     @Override
     @Deprecated
     public MailboxLock lock(boolean write) {
-    	if (write) {
+        if (write || this.mailbox.requiresWriteLock()) {
             return writeLock();
         }
         return readLock();
@@ -52,6 +61,235 @@ public class LocalMailboxLockFactory implements MailboxLockFactory {
 
     @Override
     public void close() throws Exception {
-    	
+        // noop
+    }
+
+    /**
+     * {@link MailboxLock} is a replacement of the implicit monitor lock using {@code synchronized} methods or statements on
+     * a mailbox instance. This gives extended capabilities such as timeout and limit on number of threads waiting for a
+     * particular mailbox lock. It is no longer legal to synchronize on a mailbox, otherwise an assertion error will be
+     * thrown. {@code Mailbox.beginTransaction()}) internally acquires the mailbox lock and it's released by
+     * {@code Mailbox.endTransaction()}, so that you don't have to explicitly call {@link #lock()} and {@link #close()}
+     * wrapping a mailbox transaction.
+     */
+    public class LocalMailboxLock implements MailboxLock {
+        private boolean write;
+
+        LocalMailboxLock(final boolean write) {
+            this.write = write;
+        }
+
+        private void acquireDistributedLock() throws ServiceException {
+            //TODO: consider read/write distributed lock
+            if (dLock != null && getHoldCount() == 1) {
+                try {
+                    dLock.acquire(LC.zimbra_mailbox_lock_timeout.intValue(), TimeUnit.SECONDS);
+                } catch (Exception e) {
+                    throw new LockFailedException("could not acquire distributed lock", e);
+                }
+            }
+        }
+
+        private void releaseDistributedLock() {
+            //TODO: consider read/write distributed lock
+            if (dLock != null && getHoldCount() == 1) {
+                try {
+                    dLock.release();
+                } catch (Exception e) {
+                    ZimbraLog.mailbox.warn("error while releasing distributed lock", e);
+                }
+            }
+        }
+
+        @Override
+        public int getHoldCount() {
+            return zLock.getReadHoldCount() + zLock.getWriteHoldCount();
+        }
+
+        @Override
+        public boolean isWriteLock() {
+            return isWriteLockedByCurrentThread();
+        }
+
+        @Override
+        public boolean isWriteLockedByCurrentThread() {
+            return zLock.isWriteLockedByCurrentThread();
+        }
+
+        @Override
+        public boolean isUnlocked() {
+            return !isWriteLockedByCurrentThread() && zLock.getReadHoldCount() == 0;
+        }
+
+        private synchronized boolean neverReadBeforeWrite() {
+            if (zLock.getWriteHoldCount() == 0) {
+                if (write) {
+                    Boolean readLock = assertReadLocks.get();
+                    if (readLock != null) {
+                        ZimbraLog.mailbox.error("read lock held before write", new Exception());
+                        assert(false);
+                    }
+                } else {
+                    assertReadLocks.set(true);
+                }
+            }
+            return true;
+        }
+
+        private synchronized boolean debugReleaseReadLock() {
+            //remove read lock
+            if (zLock.getReadHoldCount() == 0) {
+                assertReadLocks.remove();
+            }
+            return true;
+        }
+
+        @VisibleForTesting
+        int getQueueLength() {
+            return zLock.getQueueLength();
+        }
+
+        @VisibleForTesting
+        boolean hasQueuedThreads() {
+            return zLock.hasQueuedThreads();
+        }
+
+        /**
+         * Acquires the lock.
+         *
+         * @throws LockFailedException failed to lock
+         */
+        @Override
+        public void lock() {
+            write = write || mailbox.requiresWriteLock();
+            ZimbraLog.mailbox.trace("LOCK %s", (write ? "WRITE" : "READ"));
+            assert(neverReadBeforeWrite());
+            try {
+                if (tryLock()) {
+                    if (mailbox.requiresWriteLock() && !isWriteLockedByCurrentThread()) {
+                        //writer finished a purge while we waited
+                        promote();
+                        return;
+                    }
+                    lockStack.push(write);
+                    try {
+                        acquireDistributedLock();
+                    } catch (ServiceException e) {
+                        LockFailedException lfe = new LockFailedException("lockdb");
+                        lfe.logStackTrace();
+                        throw lfe;
+                    }
+                    return;
+                }
+                int queueLength = zLock.getQueueLength();
+                if (queueLength >= LC.zimbra_mailbox_lock_max_waiting_threads.intValue()) {
+                    // Too many threads are already waiting for the lock, can't let you queued. We don't want to log stack trace
+                    // here because once requests back up, each new incoming request falls into here, which creates too much
+                    // noise in the logs. Unless debug switch is enabled
+                    LockFailedException e = new LockFailedException("too many waiters: " + queueLength);
+                    if (DebugConfig.debugMailboxLock) {
+                        e.logStackTrace();
+                    }
+                    throw e;
+                }
+                // Wait for the lock up to the timeout.
+                if (tryLockWithTimeout()) {
+                    if (mailbox.requiresWriteLock() && !isWriteLockedByCurrentThread()) {
+                        //writer finished a purge while we waited
+                        promote();
+                        return;
+                    }
+                    lockStack.push(write);
+                    try {
+                        acquireDistributedLock();
+                    } catch (ServiceException e) {
+                        LockFailedException lfe = new LockFailedException("lockdb");
+                        lfe.logStackTrace();
+                        throw lfe;
+                    }
+                    return;
+                }
+                LockFailedException e = new LockFailedException("timeout");
+                e.logStackTrace();
+                throw e;
+            } catch (InterruptedException e) {
+                throw new LockFailedException("interrupted", e);
+            } finally {
+                assert(!isUnlocked() || debugReleaseReadLock());
+            }
+        }
+
+        private boolean tryLock() throws InterruptedException {
+            if (write) {
+                return zLock.writeLock().tryLock(0, TimeUnit.SECONDS);
+            } else {
+                return zLock.readLock().tryLock(0, TimeUnit.SECONDS);
+            }
+        }
+
+        private boolean tryLockWithTimeout() throws InterruptedException {
+            if (write) {
+                return zLock.writeLock().tryLock(LC.zimbra_mailbox_lock_timeout.intValue(), TimeUnit.SECONDS);
+            } else {
+                return zLock.readLock().tryLock(LC.zimbra_mailbox_lock_timeout.intValue(), TimeUnit.SECONDS);
+            }
+        }
+
+        @Override
+        public void close() {
+            try {
+                write = lockStack.pop();
+            } catch (EmptyStackException ese) {
+                //should only occur if locked failed; i.e. tryLock() returned error
+                //or if call site has unbalanced lock/release
+                ZimbraLog.mailbox.trace("release when not locked?");
+                assert(getHoldCount() == 0);
+                assert(debugReleaseReadLock());
+                return;
+            }
+            //keep release in order so caller doesn't have to manage write/read flag
+            ZimbraLog.mailbox.trace("RELEASE %s", (write ? "WRITE" : "READ"));
+
+            releaseDistributedLock();
+            if (write) {
+                assert(zLock.getWriteHoldCount() > 0);
+                zLock.writeLock().unlock();
+            } else {
+                zLock.readLock().unlock();
+                assert(debugReleaseReadLock());
+            }
+        }
+
+        private void promote() {
+            assert(getHoldCount() == zLock.getReadHoldCount());
+            int count = zLock.getReadHoldCount();
+            for (int i = 0; i < count - 1; i++) {
+                close();
+            }
+            zLock.readLock().unlock();
+            assert(debugReleaseReadLock());
+            for (int i = 0; i < count; i++) {
+                write = true;
+                lock();
+            }
+        }
+    }
+
+    public class LockFailedException extends RuntimeException {
+        private static final long serialVersionUID = -6899718561860023270L;
+
+        private LockFailedException(final String message) {
+            super(message);
+        }
+
+        private LockFailedException(final String message, final Throwable cause) {
+            super(message, cause);
+        }
+
+        private void logStackTrace() {
+            StringBuilder out = new StringBuilder("Failed to lock mailbox\n");
+            zLock.printStackTrace(out);
+            ZimbraLog.mailbox.error(out, this);
+        }
     }
 }
